+++ bindings/core/src/method_handler/utils.rs
@@ -18,7 +18,7 @@ use crate::{method::UtilsMethod, response::Response, Result};
 /// Call a utils method.
 pub(crate) fn call_utils_method_internal(method: UtilsMethod) -> Result<Response> {
     let response = match method {
-        UtilsMethod::Bech32ToHex { bech32 } => Response::Bech32ToHex(Client::bech32_to_hex(&bech32)?),
+        UtilsMethod::Bech32ToHex { bech32 } => Response::Bech32ToHex(Client::bech32_to_hex(bech32)?),
         UtilsMethod::HexToBech32 { hex, bech32_hrp } => Response::Bech32Address(hex_to_bech32(&hex, bech32_hrp)?),
         UtilsMethod::AliasIdToBech32 { alias_id, bech32_hrp } => {
             Response::Bech32Address(alias_id.to_bech32(bech32_hrp))