+++ sdk/examples/client/offline_signing/1_transaction_preparation.rs
@@ -52,7 +52,7 @@ async fn main() -> Result<()> {
         transaction_builder = transaction_builder.with_input(input)?;
     }
     let prepared_transaction = transaction_builder
-        .with_output(&Bech32Address::try_from_str(address)?, amount)
+        .with_output(address, amount)
         .await?
         .prepare_transaction()
         .await?;