+++ sdk/examples/client/output/nft.rs
@@ -38,8 +38,8 @@ async fn main() -> Result<()> {
 
     let token_supply = client.get_token_supply().await?;
 
-    let address = &client.get_addresses(&secret_manager).with_range(0..1).finish().await?[0];
-    request_funds_from_faucet(&faucet_url, address).await?;
+    let address = client.get_addresses(&secret_manager).with_range(0..1).finish().await?[0];
+    request_funds_from_faucet(&faucet_url, &address).await?;
     tokio::time::sleep(std::time::Duration::from_secs(20)).await;
 
     //////////////////////////////////