+++ sdk/src/types/block/address/mod.rs
@@ -10,7 +10,7 @@ use derive_more::From;
 
 pub use self::{
     alias::AliasAddress,
-    bech32::{Bech32Address, Hrp},
+    bech32::{Bech32Address, Bech32AddressLike, Hrp, HrpLike},
     ed25519::Ed25519Address,
     nft::NftAddress,
 };
@@ -113,9 +113,9 @@ impl Address {
         Bech32Address::try_from_str(address).map(|res| res.inner)
     }
 
-    /// Try to encode this address to a bech32 string with the given string Human Readable Part as prefix.
-    pub fn try_to_bech32<T: AsRef<str>>(&self, hrp: T) -> Result<Bech32Address, Error> {
-        Bech32Address::try_new(hrp.as_ref(), self)
+    /// Try to encode this address to a bech32 string with the given Human Readable Part as prefix.
+    pub fn try_to_bech32<T: HrpLike>(&self, hrp: T) -> Result<Bech32Address, Error> {
+        Bech32Address::try_new(hrp, self)
     }
 
     /// Encodes this address to a bech32 string with the given Human Readable Part as prefix.
@@ -123,10 +123,10 @@ impl Address {
         Bech32Address::new(hrp, self)
     }
 
-    /// Encodes this address to a bech32 string with the given string Human Readable Part as prefix without checking
+    /// Encodes this address to a bech32 string with the given Human Readable Part as prefix without checking
     /// validity.
-    pub fn to_bech32_unchecked<T: AsRef<str>>(&self, hrp: T) -> Bech32Address {
-        Bech32Address::new(Hrp::from_str_unchecked(hrp.as_ref()), self)
+    pub fn to_bech32_unchecked<T: HrpLike>(&self, hrp: T) -> Bech32Address {
+        Bech32Address::new(hrp.to_hrp_unchecked(), self)
     }
 
     /// Checks if an string is a valid bech32 encoded address.