+++ sdk/tests/wallet/account_recovery.rs
@@ -67,7 +67,7 @@ async fn account_recovery_with_balance_and_empty_addresses() -> Result<()> {
 
     let secret_manager = SecretManager::Mnemonic(MnemonicSecretManager::try_from_mnemonic(&mnemonic)?);
 
-    let address = &client
+    let address = client
         .get_addresses(&secret_manager)
         .with_coin_type(SHIMMER_COIN_TYPE)
         .with_bech32_hrp(client.get_bech32_hrp().await?)
@@ -77,7 +77,7 @@ async fn account_recovery_with_balance_and_empty_addresses() -> Result<()> {
         .await?[0];
 
     // Add funds to the address with account index 2 and address key_index 2, so recover works
-    iota_sdk::client::request_funds_from_faucet(crate::wallet::common::FAUCET_URL, address).await?;
+    iota_sdk::client::request_funds_from_faucet(crate::wallet::common::FAUCET_URL, &address).await?;
 
     // Wait for faucet transaction
     tokio::time::sleep(Duration::new(10, 0)).await;